+++ AffindaAPI/AffindaAPI/AffindaAPIRestClient.cs
@@ -107,7 +107,7 @@ namespace Affinda.API
             }
         }
 
-        internal HttpMessage CreateCreateResumeRequest(Stream file, string url, string data, string identifier, string fileName, string wait, string rejectDuplicates, string language, string expiryTime, string regionBias)
+        internal HttpMessage CreateCreateResumeRequest(Stream file, string url, string data, string identifier, string fileName, string wait, string rejectDuplicates, string language, string expiryTime, string regionBias, string lowPriority)
         {
             var message = _pipeline.CreateMessage();
             var request = message.Request;
@@ -161,6 +161,10 @@ namespace Affinda.API
             {
                 content.Add(new StringRequestContent(regionBias), "regionBias", null);
             }
+            if (lowPriority != null)
+            {
+                content.Add(new StringRequestContent(lowPriority), "lowPriority", null);
+            }
             content.ApplyToRequest(request);
             return message;
         }
@@ -176,14 +180,15 @@ namespace Affinda.API
         /// <param name="language"> The String to use. </param>
         /// <param name="expiryTime"> The String to use. </param>
         /// <param name="regionBias"> The String to use. </param>
+        /// <param name="lowPriority"> The String to use. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks>
         /// Uploads a resume for parsing. When successful, returns an `identifier` in the response for subsequent use with the [/resumes/{identifier}](#get-/resumes/-identifier-) endpoint to check processing status and retrieve results.&lt;br/&gt;
         /// Resumes can be uploaded as a file or a URL. In addition, data can be added directly if users want to upload directly without parsing any resume file. For uploading resume data, the `data` argument provided must be a JSON-encoded string. Data uploads will not impact upon parsing credits.
         /// </remarks>
-        public async Task<Response<Resume>> CreateResumeAsync(Stream file = null, string url = null, string data = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, CancellationToken cancellationToken = default)
+        public async Task<Response<Resume>> CreateResumeAsync(Stream file = null, string url = null, string data = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, string lowPriority = null, CancellationToken cancellationToken = default)
         {
-            using var message = CreateCreateResumeRequest(file, url, data, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias);
+            using var message = CreateCreateResumeRequest(file, url, data, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias, lowPriority);
             await _pipeline.SendAsync(message, cancellationToken).ConfigureAwait(false);
             switch (message.Response.Status)
             {
@@ -211,14 +216,15 @@ namespace Affinda.API
         /// <param name="language"> The String to use. </param>
         /// <param name="expiryTime"> The String to use. </param>
         /// <param name="regionBias"> The String to use. </param>
+        /// <param name="lowPriority"> The String to use. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks>
         /// Uploads a resume for parsing. When successful, returns an `identifier` in the response for subsequent use with the [/resumes/{identifier}](#get-/resumes/-identifier-) endpoint to check processing status and retrieve results.&lt;br/&gt;
         /// Resumes can be uploaded as a file or a URL. In addition, data can be added directly if users want to upload directly without parsing any resume file. For uploading resume data, the `data` argument provided must be a JSON-encoded string. Data uploads will not impact upon parsing credits.
         /// </remarks>
-        public Response<Resume> CreateResume(Stream file = null, string url = null, string data = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, CancellationToken cancellationToken = default)
+        public Response<Resume> CreateResume(Stream file = null, string url = null, string data = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, string lowPriority = null, CancellationToken cancellationToken = default)
         {
-            using var message = CreateCreateResumeRequest(file, url, data, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias);
+            using var message = CreateCreateResumeRequest(file, url, data, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias, lowPriority);
             _pipeline.Send(message, cancellationToken);
             switch (message.Response.Status)
             {
@@ -920,7 +926,7 @@ namespace Affinda.API
             }
         }
 
-        internal HttpMessage CreateCreateInvoiceRequest(Stream file, string url, string identifier, string fileName, string wait, string rejectDuplicates, string language, string expiryTime, string regionBias)
+        internal HttpMessage CreateCreateInvoiceRequest(Stream file, string url, string identifier, string fileName, string wait, string rejectDuplicates, string language, string expiryTime, string regionBias, string lowPriority)
         {
             var message = _pipeline.CreateMessage();
             var request = message.Request;
@@ -970,6 +976,10 @@ namespace Affinda.API
             {
                 content.Add(new StringRequestContent(regionBias), "regionBias", null);
             }
+            if (lowPriority != null)
+            {
+                content.Add(new StringRequestContent(lowPriority), "lowPriority", null);
+            }
             content.ApplyToRequest(request);
             return message;
         }
@@ -984,14 +994,15 @@ namespace Affinda.API
         /// <param name="language"> The String to use. </param>
         /// <param name="expiryTime"> The String to use. </param>
         /// <param name="regionBias"> The String to use. </param>
+        /// <param name="lowPriority"> The String to use. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks>
         /// Uploads an invoice for parsing.
         /// When successful, returns an `identifier` in the response for subsequent use with the [/invoices/{identifier}](#get-/invoices/-identifier-) endpoint to check processing status and retrieve results.
         /// </remarks>
-        public async Task<Response<Invoice>> CreateInvoiceAsync(Stream file = null, string url = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, CancellationToken cancellationToken = default)
+        public async Task<Response<Invoice>> CreateInvoiceAsync(Stream file = null, string url = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, string lowPriority = null, CancellationToken cancellationToken = default)
         {
-            using var message = CreateCreateInvoiceRequest(file, url, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias);
+            using var message = CreateCreateInvoiceRequest(file, url, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias, lowPriority);
             await _pipeline.SendAsync(message, cancellationToken).ConfigureAwait(false);
             switch (message.Response.Status)
             {
@@ -1018,14 +1029,15 @@ namespace Affinda.API
         /// <param name="language"> The String to use. </param>
         /// <param name="expiryTime"> The String to use. </param>
         /// <param name="regionBias"> The String to use. </param>
+        /// <param name="lowPriority"> The String to use. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks>
         /// Uploads an invoice for parsing.
         /// When successful, returns an `identifier` in the response for subsequent use with the [/invoices/{identifier}](#get-/invoices/-identifier-) endpoint to check processing status and retrieve results.
         /// </remarks>
-        public Response<Invoice> CreateInvoice(Stream file = null, string url = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, CancellationToken cancellationToken = default)
+        public Response<Invoice> CreateInvoice(Stream file = null, string url = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, string lowPriority = null, CancellationToken cancellationToken = default)
         {
-            using var message = CreateCreateInvoiceRequest(file, url, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias);
+            using var message = CreateCreateInvoiceRequest(file, url, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias, lowPriority);
             _pipeline.Send(message, cancellationToken);
             switch (message.Response.Status)
             {
@@ -1251,7 +1263,7 @@ namespace Affinda.API
             }
         }
 
-        internal HttpMessage CreateCreateJobDescriptionRequest(Stream file, string url, string data, string identifier, string fileName, string wait, string rejectDuplicates, string language, string expiryTime, string regionBias)
+        internal HttpMessage CreateCreateJobDescriptionRequest(Stream file, string url, string data, string identifier, string fileName, string wait, string rejectDuplicates, string language, string expiryTime, string regionBias, string lowPriority)
         {
             var message = _pipeline.CreateMessage();
             var request = message.Request;
@@ -1305,6 +1317,10 @@ namespace Affinda.API
             {
                 content.Add(new StringRequestContent(regionBias), "regionBias", null);
             }
+            if (lowPriority != null)
+            {
+                content.Add(new StringRequestContent(lowPriority), "lowPriority", null);
+            }
             content.ApplyToRequest(request);
             return message;
         }
@@ -1320,15 +1336,16 @@ namespace Affinda.API
         /// <param name="language"> The String to use. </param>
         /// <param name="expiryTime"> The String to use. </param>
         /// <param name="regionBias"> The String to use. </param>
+        /// <param name="lowPriority"> The String to use. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks>
         /// Uploads a job description for parsing.
         /// When successful, returns an `identifier` in the response for subsequent use with the [/job_descriptions/{identifier}](#get-/job_descriptions/-identifier-) endpoint to check processing status and retrieve results.
         /// Job Descriptions can be uploaded as a file or a URL. In addition, data can be added directly if users want to upload directly without parsing any resume file. For uploading resume data, the `data` argument provided must be a JSON-encoded string. Data uploads will not impact upon parsing credits.
         /// </remarks>
-        public async Task<Response<JobDescription>> CreateJobDescriptionAsync(Stream file = null, string url = null, string data = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, CancellationToken cancellationToken = default)
+        public async Task<Response<JobDescription>> CreateJobDescriptionAsync(Stream file = null, string url = null, string data = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, string lowPriority = null, CancellationToken cancellationToken = default)
         {
-            using var message = CreateCreateJobDescriptionRequest(file, url, data, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias);
+            using var message = CreateCreateJobDescriptionRequest(file, url, data, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias, lowPriority);
             await _pipeline.SendAsync(message, cancellationToken).ConfigureAwait(false);
             switch (message.Response.Status)
             {
@@ -1356,15 +1373,16 @@ namespace Affinda.API
         /// <param name="language"> The String to use. </param>
         /// <param name="expiryTime"> The String to use. </param>
         /// <param name="regionBias"> The String to use. </param>
+        /// <param name="lowPriority"> The String to use. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks>
         /// Uploads a job description for parsing.
         /// When successful, returns an `identifier` in the response for subsequent use with the [/job_descriptions/{identifier}](#get-/job_descriptions/-identifier-) endpoint to check processing status and retrieve results.
         /// Job Descriptions can be uploaded as a file or a URL. In addition, data can be added directly if users want to upload directly without parsing any resume file. For uploading resume data, the `data` argument provided must be a JSON-encoded string. Data uploads will not impact upon parsing credits.
         /// </remarks>
-        public Response<JobDescription> CreateJobDescription(Stream file = null, string url = null, string data = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, CancellationToken cancellationToken = default)
+        public Response<JobDescription> CreateJobDescription(Stream file = null, string url = null, string data = null, string identifier = null, string fileName = null, string wait = null, string rejectDuplicates = null, string language = null, string expiryTime = null, string regionBias = null, string lowPriority = null, CancellationToken cancellationToken = default)
         {
-            using var message = CreateCreateJobDescriptionRequest(file, url, data, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias);
+            using var message = CreateCreateJobDescriptionRequest(file, url, data, identifier, fileName, wait, rejectDuplicates, language, expiryTime, regionBias, lowPriority);
             _pipeline.Send(message, cancellationToken);
             switch (message.Response.Status)
             {