+++ AffindaAPI/AffindaAPI/Models/JobDescriptionRequestBody.cs
@@ -37,5 +37,7 @@ namespace Affinda.API.Models
         public string ExpiryTime { get; }
         /// <summary> A JSON representation of the RegionBias object. </summary>
         public string RegionBias { get; }
+        /// <summary> Explicitly mark this document as low priority. </summary>
+        public string LowPriority { get; }
     }
 }